Move common components to shared folder
