+++ tps-monitoring/src/csv-report/tps_stats.csv
@@ -1,81 +1,45 @@
 block,timestamp,total_transactions,our_transactions,total_tps,our_tps
-15220,2025-07-21T15:29:40.032Z,0,0,0.00,0.00
-15221,2025-07-21T15:29:40.115Z,0,0,0.00,0.00
-15222,2025-07-21T15:29:40.215Z,0,0,0.00,0.00
-15223,2025-07-21T15:29:40.317Z,0,0,0.00,0.00
-15224,2025-07-21T15:29:40.416Z,0,0,0.00,0.00
-15225,2025-07-21T15:29:40.512Z,0,0,0.00,0.00
-15226,2025-07-21T15:29:40.610Z,0,0,0.00,0.00
-15227,2025-07-21T15:29:40.713Z,0,0,0.00,0.00
-15228,2025-07-21T15:29:40.810Z,0,0,0.00,0.00
-15229,2025-07-21T15:29:40.917Z,0,0,0.00,0.00
-15230,2025-07-21T15:29:41.018Z,0,0,0.00,0.00
-15231,2025-07-21T15:29:41.112Z,0,0,0.00,0.00
-15232,2025-07-21T15:29:41.212Z,0,0,0.00,0.00
-15233,2025-07-21T15:29:41.311Z,0,0,0.00,0.00
-15234,2025-07-21T15:29:41.412Z,0,0,0.00,0.00
-15235,2025-07-21T15:29:41.513Z,0,0,0.00,0.00
-15236,2025-07-21T15:29:41.612Z,0,0,0.00,0.00
-15237,2025-07-21T15:29:41.712Z,0,0,0.00,0.00
-15238,2025-07-21T15:29:41.815Z,0,0,0.00,0.00
-15239,2025-07-21T15:29:41.911Z,0,0,0.00,0.00
-15240,2025-07-21T15:29:42.015Z,0,0,0.00,0.00
-15241,2025-07-21T15:29:42.112Z,0,0,0.00,0.00
-15242,2025-07-21T15:29:42.219Z,0,0,0.00,0.00
-15243,2025-07-21T15:29:42.312Z,0,0,0.00,0.00
-15244,2025-07-21T15:29:42.415Z,0,0,0.00,0.00
-15245,2025-07-21T15:29:42.515Z,0,0,0.00,0.00
-15246,2025-07-21T15:29:42.611Z,0,0,0.00,0.00
-15247,2025-07-21T15:29:42.710Z,0,0,0.00,0.00
-15248,2025-07-21T15:29:42.816Z,0,0,0.00,0.00
-15249,2025-07-21T15:29:42.911Z,0,0,0.00,0.00
-15250,2025-07-21T15:29:43.008Z,0,0,0.00,0.00
-15251,2025-07-21T15:29:43.111Z,0,0,0.00,0.00
-15252,2025-07-21T15:29:43.212Z,0,0,0.00,0.00
-15253,2025-07-21T15:29:43.313Z,0,0,0.00,0.00
-15254,2025-07-21T15:29:43.417Z,0,0,0.00,0.00
-15255,2025-07-21T15:29:43.511Z,0,0,0.00,0.00
-15256,2025-07-21T15:29:43.609Z,0,0,0.00,0.00
-15257,2025-07-21T15:29:43.712Z,0,0,0.00,0.00
-15258,2025-07-21T15:29:43.810Z,0,0,0.00,0.00
-15259,2025-07-21T15:29:43.909Z,0,0,0.00,0.00
-15260,2025-07-21T15:29:44.012Z,0,0,0.00,0.00
-15261,2025-07-21T15:29:44.110Z,0,0,0.00,0.00
-15262,2025-07-21T15:29:44.210Z,0,0,0.00,0.00
-15263,2025-07-21T15:29:44.314Z,0,0,0.00,0.00
-15264,2025-07-21T15:29:44.410Z,0,0,0.00,0.00
-15265,2025-07-21T15:29:44.515Z,0,0,0.00,0.00
-15266,2025-07-21T15:29:44.611Z,0,0,0.00,0.00
-15267,2025-07-21T15:29:44.710Z,0,0,0.00,0.00
-15268,2025-07-21T15:29:44.818Z,0,0,0.00,0.00
-15269,2025-07-21T15:29:44.909Z,0,0,0.00,0.00
-15270,2025-07-21T15:29:45.019Z,0,0,0.00,0.00
-15271,2025-07-21T15:29:45.109Z,0,0,0.00,0.00
-15272,2025-07-21T15:29:45.210Z,0,0,0.00,0.00
-15273,2025-07-21T15:29:45.309Z,0,0,0.00,0.00
-15274,2025-07-21T15:29:45.413Z,0,0,0.00,0.00
-15275,2025-07-21T15:29:45.515Z,0,0,0.00,0.00
-15276,2025-07-21T15:29:45.611Z,0,0,0.00,0.00
-15277,2025-07-21T15:29:45.710Z,0,0,0.00,0.00
-15278,2025-07-21T15:29:45.813Z,0,0,0.00,0.00
-15279,2025-07-21T15:29:45.915Z,0,0,0.00,0.00
-15280,2025-07-21T15:29:46.016Z,0,0,0.00,0.00
-15281,2025-07-21T15:29:46.110Z,0,0,0.00,0.00
-15282,2025-07-21T15:29:46.215Z,0,0,0.00,0.00
-15283,2025-07-21T15:29:46.312Z,0,0,0.00,0.00
-15284,2025-07-21T15:29:46.416Z,0,0,0.00,0.00
-15285,2025-07-21T15:29:46.512Z,0,0,0.00,0.00
-15286,2025-07-21T15:29:46.613Z,0,0,0.00,0.00
-15287,2025-07-21T15:29:46.718Z,0,0,0.00,0.00
-15288,2025-07-21T15:29:46.812Z,0,0,0.00,0.00
-15289,2025-07-21T15:29:46.914Z,0,0,0.00,0.00
-15290,2025-07-21T15:29:47.014Z,0,0,0.00,0.00
-15291,2025-07-21T15:29:47.114Z,0,0,0.00,0.00
-15292,2025-07-21T15:29:47.216Z,0,0,0.00,0.00
-15293,2025-07-21T15:29:47.314Z,0,0,0.00,0.00
-15294,2025-07-21T15:29:47.415Z,0,0,0.00,0.00
-15295,2025-07-21T15:29:47.511Z,0,0,0.00,0.00
-15296,2025-07-21T15:29:47.615Z,0,0,0.00,0.00
-15297,2025-07-21T15:29:47.714Z,0,0,0.00,0.00
-15298,2025-07-21T15:29:47.813Z,0,0,0.00,0.00
-15299,2025-07-21T15:29:47.909Z,0,0,0.00,0.00
+136,2025-07-22T05:46:56.709Z,0,0,0.00,0.00
+137,2025-07-22T05:46:56.713Z,0,0,0.00,0.00
+138,2025-07-22T05:46:56.813Z,0,0,0.00,0.00
+139,2025-07-22T05:46:56.915Z,0,0,0.00,0.00
+140,2025-07-22T05:46:57.011Z,0,0,0.00,0.00
+141,2025-07-22T05:46:57.110Z,0,0,0.00,0.00
+142,2025-07-22T05:46:57.207Z,0,0,0.00,0.00
+143,2025-07-22T05:46:57.315Z,0,0,0.00,0.00
+144,2025-07-22T05:46:57.408Z,0,0,0.00,0.00
+145,2025-07-22T05:46:57.512Z,0,0,0.00,0.00
+146,2025-07-22T05:46:57.612Z,0,0,0.00,0.00
+147,2025-07-22T05:46:57.715Z,0,0,0.00,0.00
+148,2025-07-22T05:46:57.816Z,0,0,0.00,0.00
+149,2025-07-22T05:46:57.915Z,0,0,0.00,0.00
+150,2025-07-22T05:46:58.009Z,0,0,0.00,0.00
+151,2025-07-22T05:46:58.110Z,0,0,0.00,0.00
+152,2025-07-22T05:46:58.215Z,0,0,0.00,0.00
+153,2025-07-22T05:46:58.315Z,0,0,0.00,0.00
+154,2025-07-22T05:46:58.415Z,0,0,0.00,0.00
+155,2025-07-22T05:46:58.512Z,0,0,0.00,0.00
+156,2025-07-22T05:46:58.612Z,0,0,0.00,0.00
+157,2025-07-22T05:46:58.714Z,0,0,0.00,0.00
+158,2025-07-22T05:46:58.814Z,0,0,0.00,0.00
+159,2025-07-22T05:46:58.908Z,0,0,0.00,0.00
+160,2025-07-22T05:46:59.008Z,0,0,0.00,0.00
+161,2025-07-22T05:46:59.111Z,0,0,0.00,0.00
+162,2025-07-22T05:46:59.215Z,0,0,0.00,0.00
+163,2025-07-22T05:46:59.315Z,0,0,0.00,0.00
+164,2025-07-22T05:46:59.414Z,0,0,0.00,0.00
+165,2025-07-22T05:46:59.512Z,0,0,0.00,0.00
+166,2025-07-22T05:46:59.615Z,0,0,0.00,0.00
+167,2025-07-22T05:46:59.711Z,0,0,0.00,0.00
+168,2025-07-22T05:46:59.809Z,0,0,0.00,0.00
+169,2025-07-22T05:46:59.916Z,0,0,0.00,0.00
+170,2025-07-22T05:47:00.012Z,0,0,0.00,0.00
+171,2025-07-22T05:47:00.111Z,0,0,0.00,0.00
+172,2025-07-22T05:47:00.214Z,0,0,0.00,0.00
+173,2025-07-22T05:47:00.313Z,0,0,0.00,0.00
+174,2025-07-22T05:47:00.412Z,0,0,0.00,0.00
+175,2025-07-22T05:47:00.515Z,0,0,0.00,0.00
+176,2025-07-22T05:47:00.611Z,0,0,0.00,0.00
+177,2025-07-22T05:47:00.715Z,0,0,0.00,0.00
+178,2025-07-22T05:47:00.811Z,0,0,0.00,0.00
+179,2025-07-22T05:47:00.912Z,0,0,0.00,0.00
